Removed and added
Removed unwanted classes and methods
Changed . We will lanuch the browser based properties file datas
