--- src/main/java/Enum/Browser_Enum.java
@@ -1,10 +1,0 @@
-package Enum;
-
-public enum Browser_Enum {
-
-    CHROME ,
-    EDGE ,
-    OPERA ,
-    SAFARI ,
-    FIREBOX;
-}
