--- src/main/java/base_class/dri.java
@@ -1,8 +1,0 @@
-package base_class;
-
-public class dri {
-	public static void main(String[] args) {
-		System.out.println("hi");
-	}
-
-}
