--- src/main/java/interfaces/sleeps.java
@@ -1,6 +1,0 @@
-package interfaces;
-
-public interface sleeps {
-
-	public void sleep(int secondes);
-}
